+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Change: Note compatibility through WP 5.4+
 * Change: Update links to coffee2code.com to be HTTPS
 * Unit tests:
+    * New: Add tests for `display_time()`
     * Change: Use HTTPS for link to WP SVN repository in bin script for configuring unit tests (and delete commented-out code)
     * Change: Remove unnecessary unregistering of hooks in `tearDown()`